pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "xminer"
version = "0.1"
description = "Data collector for X profiles"
# readme = "README.md"
requires-python = ">=3.11"
authors = [
    { name = "Your Name" },
]
dependencies = [
    "pandas==2.3.1",
    "numpy==2.3.2",
    "tweepy==4.16.0",
    "SQLAlchemy==2.0.34",
    "psycopg2-binary==2.9.11",
    "python-dotenv==1.0.1",
    "tenacity==8.5.0",
    "pyyaml==6.0.2",
    "pyarrow>=15",
    "typer",
    "ipykernel",
    "plotly",
    "openpyxl",
    "requests",
    "beautifulsoup4",
    "geopy>=2.4.0"
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
